requirements.txt
google-api-python-client
pandas
tqdm
vaderSentiment
nltk
jsonify
google-generativeai